# docs/swagger.yaml
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.GenericResponse:
    properties:
      data:
        type: string
      success:
        type: boolean
    type: object
  model.Player:
    properties:
      createdAt:
        type: string
      draws:
        type: integer
      elo:
        type: integer
      id:
        type: integer
      isPlaying:
        type: boolean
      lastPlayedAt:
        type: string
      losses:
        type: integer
      rate:
        type: number
      username:
        type: string
      wins:
        type: integer
    type: object
  model.PlayerListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Player'
        type: array
      resultCount:
        type: integer
      totalCount:
        type: integer
    type: object
  model.PlayerRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: lukamatosevic5@gmail.com
    name: Luka Matošević
    url: https://lukamatosevic.com
  license:
    name: MIT 2023
    url: https://www.mit.edu/~amini/LICENSE.md
paths:
...
  /v1/players:
    get:
      description: Query and list players
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Filter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.PlayerListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Query and list players
      tags:
      - players
  /v1/players/{id}:
    get:
      description: Find one player
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Find one player
      tags:
      - players
  /v1/players/delete:
    delete:
      consumes:
      - application/json
      description: Delete player account
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete player account
      tags:
      - players
  /v1/players/register:
    post:
      consumes:
      - application/json
      description: Register new player
      parameters:
      - description: Register player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/model.PlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Register new player
      tags:
      - players
  /v1/players/update:
    put:
      consumes:
      - application/json
      description: Update player account
      parameters:
      - description: Update player data
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/model.PlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update player account
      tags:
      - players
securityDefinitions:
  ApiKeyAuth:
    description: The access token obtained from /login endpoint, required for accessing
      protected routes
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"